Add the alpha release of bottle to the git repo
This represents the alpha release of gobottle/bottle.

While the tool IS being used in a production environment right now, it
does not have sufficient platform support nor documentation to even be
considered a beta software release.

Neverthless, anyone is encouraged to play around with and contribute,
and even use it in your own projects.