Improve recursive resolver, add commands for v0.2.0
This commit is an almost completed rewrite of bottle to remove
all of the HACKs and TODOs that were used to resolve package
dependencies in version 0.1.0.

Now package resolution is multi-threaded, and does not depend
on "go get" to resolve the dependencies for any package with
a Bottle.toml file.

Additionally, the default behavior is no longer "build" and
instead commands have been added that work in the same way as
the "cargo" and "go" cli tools.